+++ cli/src/commands/git/mod.rs
@@ -21,6 +21,7 @@ mod init;
 mod push;
 mod remote;
 mod root;
+mod sync;
 
 use std::io::Write as _;
 
@@ -57,6 +58,8 @@ use self::remote::RemoteCommand;
 use self::remote::cmd_git_remote;
 use self::root::GitRootArgs;
 use self::root::cmd_git_root;
+use self::sync::GitSyncArgs;
+use self::sync::cmd_git_sync;
 use crate::cli_util::CommandHelper;
 use crate::cli_util::WorkspaceCommandHelper;
 use crate::command_error::CommandError;
@@ -88,6 +91,7 @@ pub enum GitCommand {
     #[command(subcommand)]
     Remote(RemoteCommand),
     Root(GitRootArgs),
+    Sync(GitSyncArgs),
 }
 
 pub async fn cmd_git(
@@ -105,6 +109,7 @@ pub async fn cmd_git(
         GitCommand::Push(args) => cmd_git_push(ui, command, args).await,
         GitCommand::Remote(args) => cmd_git_remote(ui, command, args).await,
         GitCommand::Root(args) => cmd_git_root(ui, command, args).await,
+        GitCommand::Sync(args) => cmd_git_sync(ui, command, args).await,
     }
 }